Add the recipient balance on the front end.
Update loadBlockchainData's async function with a Map named recipientsBalance
and adds it to React state.

Iterate through each proposal and had the proposal's recipient balance
to recipientsBalance Map. Only fetch the recipient's balance if it is not already defined, to
avoid unecessary async ethers calls to the blockchain.

Update the Proposals component to show the recipient's balance, alongside
their wallet address.